removing createSession method from client-financials moving it to statement view; statement view now works but still needs some scrollbar work done
