+++ epfirms-client/src/app/client-portal/client-financials/client-financials.component.ts
@@ -75,34 +75,7 @@ export class ClientFinancialsComponent implements OnInit {
     this.matterBalances[`${matter.case_id}`] = balanceDue;
   }
 
-  createPaymentSession(matter, balance) : void {
-    let paymentData = {
-      balance: balance,
-    }
-    this.stripeService.createPaymentSession(paymentData).subscribe(res => {
-      // we receive the session url and the id
-      let sessionId = res.session_id;
-      let url = res.url;
-      console.log("URL", url);
-      // we need to add the session id to the statement to verify payment
-      // when the webhook sends back a request to the server on fufillment
-      // ideally, we don't want to navigate to the new window until this session
-      // has been saved as we don't ever want there to be a time when the session
-      // id's do not match for some reason
-      let statementCount = 0;
-      console.log(this.matterStatements[`${matter.case_id}`])
-      this.matterStatements[`${matter.case_id}`].forEach(statement => {
-        statementCount += 1;
-        statement.stripe_session_id = sessionId;
-        this.statementService.update(statement).subscribe(res => {
-          // only change the window after the last statement has been updated
-          if (statementCount == this.matterStatements[`${matter.case_id}`].length){
-            window.location.replace(url);
-          }
-        });
-      });
-    });
-  }
+
 
   openStatementView(matter) : void {
     console.log(matter);
