+++ epfirms-client/src/app/client-portal/client-financials/statement-view/statement-view.component.html
@@ -2,75 +2,84 @@
 <div class="bg-white">
     <div class="max-w-7xl mx-auto py-16 px-4 sm:px-6 lg:pb-24 lg:px-8">
       <div class="max-w-xl">
-        <h1 class="text-2xl font-extrabold tracking-tight text-gray-900 sm:text-3xl">Statements for CASE NUMBER</h1>
+        <h1 class="text-2xl font-extrabold tracking-tight text-gray-900 sm:text-3xl">Statements</h1>
         <p class="mt-2 text-sm text-gray-500">Review statements that have been issued for this case.</p>
       </div>
-  
-      <div class="mt-16">
+      
+      <div *ngFor="let statement of statements" class="mt-16">
         <h2 class="sr-only">Statements for CASE NUMBER</h2>
   
         <div class="space-y-20">
           <div>
-            <h3 class="sr-only">Order placed on <time datetime="2021-01-22">January 22, 2021</time></h3>
+            <h3 class="sr-only">Statement Generated on <time datetime="2021-01-22">{{statement.created_at}}</time></h3>
   
             <div class="bg-gray-50 rounded-lg py-6 px-4 sm:px-6 sm:flex sm:items-center sm:justify-between sm:space-x-6 lg:space-x-8">
               <dl class="divide-y divide-gray-200 space-y-6 text-sm text-gray-600 flex-auto sm:divide-y-0 sm:space-y-0 sm:grid sm:grid-cols-3 sm:gap-x-6 lg:w-1/2 lg:flex-none lg:gap-x-8">
                 <div class="flex justify-between sm:block">
-                  <dt class="font-medium text-gray-900">Date placed</dt>
+                  <dt class="font-medium text-gray-900">Date</dt>
                   <dd class="sm:mt-1">
-                    <time datetime="2021-01-22">January 22, 2021</time>
+                    <time datetime="2021-01-22">{{statement.created_at | date}}</time>
                   </dd>
                 </div>
                 <div class="flex justify-between pt-6 sm:block sm:pt-0">
-                  <dt class="font-medium text-gray-900">Order number</dt>
+                  <dt class="font-medium text-gray-900">Due Date</dt>
                   <dd class="sm:mt-1">
-                    WU88191111
+                    {{statement.due_date | date}}
                   </dd>
                 </div>
                 <div class="flex justify-between pt-6 font-medium text-gray-900 sm:block sm:pt-0">
                   <dt>Total amount</dt>
                   <dd class="sm:mt-1">
-                    $238.00
+                    {{statement.balance_due | currency}}
+                  </dd>
+                </div>
+                <div class="flex justify-between pt-6 font-medium text-gray-900 sm:block sm:pt-0">
+                  <dt>Status</dt>
+                  <dd class="sm:mt-1">
+                    {{statement.status}}
                   </dd>
                 </div>
               </dl>
-              <a href="#" class="w-full flex items-center justify-center bg-white mt-6 py-2 px-4 border border-gray-300 rounded-md shadow-sm text-sm font-medium text-gray-700 hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500 sm:w-auto sm:mt-0">
-                View Invoice
+              <a *ngIf="statement.status == 'UNPAID'" (click)="createPaymentSession(statement)" class="w-full flex items-center justify-center bg-white mt-6 py-2 px-4 border border-gray-300 rounded-md shadow-sm text-sm font-medium text-gray-700 hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500 sm:w-auto sm:mt-0">
+                Pay Now
                 <span class="sr-only">for order WU88191111</span>
               </a>
             </div>
   
             <table class="mt-4 w-full text-gray-500 sm:mt-6">
               <caption class="sr-only">
-                Products
+                Transactions
               </caption>
               <thead class="sr-only text-sm text-gray-500 text-left sm:not-sr-only">
                 <tr>
-                  <th scope="col" class="sm:w-2/5 lg:w-1/3 pr-8 py-3 font-normal">Product</th>
-                  <th scope="col" class="hidden w-1/5 pr-8 py-3 font-normal sm:table-cell">Price</th>
-                  <th scope="col" class="hidden pr-8 py-3 font-normal sm:table-cell">Status</th>
-                  <th scope="col" class="w-0 py-3 font-normal text-right">Info</th>
+                  <th scope="col" class="">Date</th>
+                  <th scope="col" class="">Staff</th>
+                  <th scope="col" class="">Service</th>
+                  <th scope="col" class="">Hours</th>
+                  <th scope="col" class="">Amount</th>
                 </tr>
               </thead>
               <tbody class="border-b border-gray-200 divide-y divide-gray-200 text-sm sm:border-t">
-                <tr>
-                  <td class="py-6 pr-8">
+                <tr *ngFor="let bill of statementBill[statement.matter_id]">
+                  <td class="py-6 pr-2">
                     <div class="flex items-center">
-                      <img src="https://tailwindui.com/img/ecommerce-images/order-history-page-02-product-01.jpg" alt="Detail of mechanical pencil tip with machined black steel shaft and chrome lead tip." class="w-16 h-16 object-center object-cover rounded mr-6">
                       <div>
-                        <div class="font-medium text-gray-900">Machined Pen and Pencil Set</div>
-                        <div class="mt-1 sm:hidden">$70.00</div>
+                        <div class="font-medium text-gray-900">{{bill.created_at | date}}</div>
+                        
                       </div>
                     </div>
                   </td>
                   <td class="hidden py-6 pr-8 sm:table-cell">
-                    $70.00
+                    {{bill.employee_name}}
                   </td>
                   <td class="hidden py-6 pr-8 sm:table-cell">
-                    Delivered Jan 25, 2021
+                    {{bill.description}}
                   </td>
-                  <td class="py-6 font-medium text-right whitespace-nowrap">
-                    <a href="#" class="text-indigo-600">View<span class="hidden lg:inline"> Product</span><span class="sr-only">, Machined Pen and Pencil Set</span></a>
+                  <td class="hidden py-6 pr-8 sm:table-cell">
+                    {{bill.hours}}
+                  </td>
+                  <td class="hidden py-6 pr-8 sm:table-cell">
+                    {{bill.amount | currency}}
                   </td>
                 </tr>
